refacto: split integration tests
